product_ordering_service/requirements.txt
django
djangorestframework
psycopg2-binary
requests
pyyaml
pytest
pytest-django
pytest-cov
model_bakery